+++ packages/html2sb/package.json
@@ -12,7 +12,7 @@
     "build": "tsc --outDir dist --project .",
     "watch": "npm run build -- --watch",
     "prepublishOnly": "rm -rf dist/* && npm run build",
-    "test": "ava -v",
+    "test": "vitest run",
     "lint": "biome check ./src ./test",
     "lint:fix": "biome check --write ./src ./test",
     "test:watch": "npm test -- --watch"
@@ -35,7 +35,8 @@
     "@biomejs/biome": "^1.9.4",
     "ava": "^3.8.1",
     "ts-node": "^8.10.1",
-    "typescript": "5.7.3"
+    "typescript": "5.7.3",
+    "vitest": "^3.0.5"
   },
   "ava": {
     "typescript": {
