+++ packages/md2sb/test/complexText.test.ts
@@ -1,6 +1,6 @@
-import test from "ava";
+import { test } from 'vitest';
 import loadAndAssert from "./helpers/loadAndAssert";
 
 ["link-includes-image", "list-strong-style-text"].forEach((type) => {
-  test("convert " + type, (t) => loadAndAssert(t, type));
+  test("convert " + type, () => loadAndAssert(type));
 });
