+++ packages/md2sb/test/helpers/loadAndAssert.ts
@@ -1,14 +1,15 @@
-import * as fs from "node:fs";
-import * as path from "node:path";
+import { expect } from 'vitest'
+import fs from "node:fs";
+import path from "node:path";
 import md2sb from "../../src/main";
 
-export default async (t, type) => {
+export default async (type) => {
   const markdown = fs.readFileSync(
     path.resolve("test/fixtures/md/" + type + ".md"),
   );
   const input = await md2sb(markdown);
-  const expect = fs
+  const expected = fs
     .readFileSync(path.resolve("test/fixtures/scrapbox/" + type + ".txt"))
     .toString();
-  t.is(input, expect);
+    expect(input).toEqual(expected);
 };
