+++ packages/md2sb/test/textDecoration.test.ts
@@ -1,4 +1,4 @@
-import test from "ava";
+import { test } from 'vitest'
 import loadAndAssert from "./helpers/loadAndAssert";
 
 [
@@ -10,5 +10,5 @@ import loadAndAssert from "./helpers/loadAndAssert";
   "image",
   "code",
 ].forEach((type) => {
-  test("convert " + type, (t) => loadAndAssert(t, type));
+  test("convert " + type, () => loadAndAssert(type));
 });
